worked on the requisition summary

+++ utils/transform.ts
@@ -77,7 +77,7 @@ export const formatOfferDate = (dateString: string) => {
   if (!dateString) return "";
   const date = new Date(dateString);
   const day = date.getDate();
-  const month = date.toLocaleDateString('en-US', { month: 'long' });
+  const month = date.toLocaleDateString('en-US', { month: 'short' });
   const year = date.getFullYear();
 
   // Logic for "st", "nd", "rd", "th"
